Moving shopping item manager to Entity\Product
