+++ src/Demo/CartBundle/Entity/Product/ShoppingItemManager.php
@@ -6,7 +6,7 @@
  * Time: 2:20 PM
  */
 
-namespace Demo\CartBundle\Entity;
+namespace Demo\CartBundle\Entity\Product;
 
 
 /**
